# docs/api/swagger.yaml
basePath: /
definitions:
...
  errors.HTTPError:
    properties:
      code:
        type: string
      detail:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the Fuse service.
  title: Fuse API
  version: 1.0.0
paths:
  /auth/{provider}:
    get:
      description: Redirects the client to the selected OAuth provider.
      parameters:
      - description: OAuth provider
        in: path
        name: provider
        required: true
        type: string
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
      summary: Start OAuth authentication
      tags:
      - authentication
  /auth/{provider}/callback:
    get:
      description: Processes the OAuth callback, sets the session cookie, and redirects
        to the frontend.
      parameters:
      - description: OAuth provider
        in: path
        name: provider
        required: true
        type: string
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
      summary: Complete OAuth authentication
      tags:
      - authentication
...
  /auth/logout:
    post:
      description: Deletes the current session and clears its cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      summary: End the current session
      tags:
      - authentication
  /auth/status:
    get:
      description: Returns the user associated with the current session cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
      summary: Get the authenticated user
      tags:
      - authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"