# internal/interfaces/server/auth/handler.go
package user

import (
	"context"
	"encoding/json"
	"net/http"
	"net/url"
	"strings"

	"fuse/internal/services/auth"
	"fuse/pkg/config"
	"fuse/pkg/errors"
	"fuse/pkg/log"

	"github.com/go-chi/chi/v5"
	"github.com/markbates/goth/gothic"
)

type Handler struct {
	authSvc *auth.Service
	cfg     *config.Config
}

const authReturnCookieName = "fuse_auth_return_to"

func NewHandler(authService *auth.Service, cfg *config.Config) *Handler {
	return &Handler{
		authSvc: authService,
		cfg:     cfg,
	}
}

func (h *Handler) RegisterRoutes(r chi.Router) {
	r.Route("/auth", func(r chi.Router) {
		r.Get("/{provider}", h.BeginAuth)
		r.Get("/{provider}/callback", h.AuthCallback)
		r.Post("/logout", h.Logout)
		r.Get("/status", h.GetAuthStatus)
	})
}

// @Summary		Start OAuth authentication
// @Description	Redirects the client to the selected OAuth provider.
// @Tags			authentication
// @Param			provider	path	string	true	"OAuth provider"
// @Success		302
// @Failure		400	{object}	errors.HTTPError
// @Router			/auth/{provider} [get]
func (h *Handler) BeginAuth(w http.ResponseWriter, r *http.Request) {
	provider := chi.URLParam(r, "provider")
	if provider == "" {
		errors.WriteError(w, errors.ErrBadRequest.WithDetail("provider is required"))
		return
	}
	if returnTo := h.validReturnURL(r.URL.Query().Get("return_to")); returnTo != "" {
		h.setReturnCookie(w, returnTo)
	}

	r = r.WithContext(context.WithValue(r.Context(), gothic.ProviderParamKey, provider))
	gothic.BeginAuthHandler(w, r)
}

// @Summary		Complete OAuth authentication
// @Description	Processes the OAuth callback, sets the session cookie, and redirects to the frontend.
// @Tags			authentication
// @Param			provider	path	string	true	"OAuth provider"
// @Success		302
// @Failure		400	{object}	errors.HTTPError
// @Failure		401	{object}	errors.HTTPError
// @Router			/auth/{provider}/callback [get]
func (h *Handler) AuthCallback(w http.ResponseWriter, r *http.Request) {
	provider := chi.URLParam(r, "provider")

	if provider == "" {
		errors.WriteError(w, errors.ErrBadRequest.WithDetail("provider is required"))
		return
	}
	r = r.WithContext(context.WithValue(r.Context(), gothic.ProviderParamKey, provider))
	gu, err := gothic.CompleteUserAuth(w, r)
	if err != nil {
		log.Warn("oauth callback error: %v", err)
		errors.WriteError(w, errors.ErrUnauthorized.WithDetail("authentication failed"))
		return
	}

	_, sid, err := h.authSvc.HandleOAuthCallback(r.Context(), gu)
	if err != nil {
		log.Warn("auth service error: %v", err)
		errors.WriteError(w, errors.ErrUnauthorized.WithDetail("authentication failed in service"))
		return
	}

	h.setSessionCookie(w, sid)

	redirectURL := h.cfg.Frontend.URL
	if cookie, cookieErr := r.Cookie(authReturnCookieName); cookieErr == nil {
		if returnTo := h.validReturnURL(cookie.Value); returnTo != "" {
			redirectURL = returnTo
		}
	}
	h.clearReturnCookie(w)

	http.Redirect(w, r, redirectURL, http.StatusFound)

}

// @Summary		Get the authenticated user
// @Description	Returns the user associated with the current session cookie.
// @Tags			authentication
// @Produce		json
// @Success		200	{object}	map[string]interface{}
// @Failure		401	{object}	errors.HTTPError
// @Router			/auth/status [get]
func (h *Handler) GetAuthStatus(w http.ResponseWriter, r *http.Request) {
	sessionID := h.getSessionFromCookie(r)
	if sessionID == "" {
		errors.WriteError(w, errors.ErrUnauthorized.WithDetail("missing session cookie"))
		return
	}

	u, err := h.authSvc.ValidateSession(r.Context(), sessionID)
	if err != nil {
		log.Warn("session validation error: %v", err)
		errors.WriteError(w, errors.ToHTTP(err).WithDetail("invalid session"))
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	if err := json.NewEncoder(w).Encode(u); err != nil {
		log.Error("failed to encode user response: %v", err)
	}
}

// @Summary		End the current session
// @Description	Deletes the current session and clears its cookie.
// @Tags			authentication
// @Produce		json
// @Success		200	{object}	map[string]string
// @Failure		401	{object}	errors.HTTPError
// @Failure		500	{object}	errors.HTTPError
// @Router			/auth/logout [post]
func (h *Handler) Logout(w http.ResponseWriter, r *http.Request) {
	sessionID := h.getSessionFromCookie(r)
	if sessionID == "" {
		errors.WriteError(w, errors.ErrUnauthorized.WithDetail("missing session cookie"))
		return
	}

	if err := h.authSvc.Logout(r.Context(), sessionID); err != nil {
		log.Warn("logout error: %v", err)
		errors.WriteError(w, errors.ErrInternalServer.WithDetail("failed to logout"))
		return
	}

	h.clearSessionCookie(w)
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{"message":"logged out successfully"}`))
}

// NOTE: Utils
func (h *Handler) setSessionCookie(w http.ResponseWriter, sessionID string) {
	sameSite := http.SameSiteLaxMode
	if h.cfg.Environment == "production" {
		sameSite = http.SameSiteStrictMode
	}

	cookie := &http.Cookie{
		Name:     h.cfg.Session.Name,
		Value:    sessionID,
		Path:     h.cfg.Session.Cookie.Path,
		Domain:   h.cfg.Session.Cookie.Domain,
		Secure:   h.cfg.Session.Cookie.Secure,
		HttpOnly: h.cfg.Session.Cookie.HTTPOnly,
		SameSite: sameSite,
		MaxAge:   h.cfg.Session.Duration,
	}
	http.SetCookie(w, cookie)
}

func (h *Handler) clearSessionCookie(w http.ResponseWriter) {
	cookie := &http.Cookie{
		Name:   h.cfg.Session.Name,
		Value:  "",
		Path:   h.cfg.Session.Cookie.Path,
		MaxAge: -1,
	}
	http.SetCookie(w, cookie)
}

func (h *Handler) getSessionFromCookie(r *http.Request) string {
	cookie, err := r.Cookie(h.cfg.Session.Name)
	if err != nil {
		return ""
	}
	return cookie.Value
}

func (h *Handler) validReturnURL(value string) string {
	value = strings.TrimSpace(value)
	if value == "" {
		return ""
	}

	returnURL, err := url.Parse(value)
	if err != nil {
		return ""
	}
	frontendURL, err := url.Parse(h.cfg.Frontend.URL)
	if err != nil {
		return ""
	}
	if returnURL.Scheme != frontendURL.Scheme || returnURL.Host != frontendURL.Host {
		return ""
	}
	if returnURL.User != nil {
		return ""
	}

	return returnURL.String()
}

func (h *Handler) setReturnCookie(w http.ResponseWriter, returnTo string) {
	http.SetCookie(w, &http.Cookie{
		Name:     authReturnCookieName,
		Value:    returnTo,
		Path:     "/auth",
		Secure:   h.cfg.Session.Cookie.Secure,
		HttpOnly: true,
		SameSite: http.SameSiteLaxMode,
		MaxAge:   600,
	})
}

func (h *Handler) clearReturnCookie(w http.ResponseWriter) {
	http.SetCookie(w, &http.Cookie{
		Name:     authReturnCookieName,
		Value:    "",
		Path:     "/auth",
		Secure:   h.cfg.Session.Cookie.Secure,
		HttpOnly: true,
		SameSite: http.SameSiteLaxMode,
		MaxAge:   -1,
	})
}